% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTs_habitat.R
\docType{data}
\name{CTs_habitat}
\alias{CTs_habitat}
\title{Example of data of coordinates and covariates of each camera trap.}
\format{
\code{CTs_habitat}
Data frame with 24 rows and 5 variables:
\describe{
\item{Camera}{name camera trap}
\item{X}{X coordinate}
\item{Y}{Y coordinate}
\item{Location}{Study area}
\item{Veg_type}{Covariate, vegetation types}
}
}
\usage{
CTs_habitat
}
\description{
Example of data of coordinates and covariates of each camera trap.
}
\keyword{datasets}
