% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTs_habitat2.R
\docType{data}
\name{CTs_habitat2}
\alias{CTs_habitat2}
\title{Example of second data of coordinates and covariates of each camera trap.}
\format{
\code{CTs_habitat2}
Data frame with 5 rows and 6 variables:
\describe{
\item{Camera}{name camera trap}
\item{X}{X coordinate}
\item{Y}{Y coordinate}
\item{Var1}{Covariable 1}
\item{Var2}{Covariable 2}
\item{Veg_type}{Covariate, vegetation types}
}
}
\usage{
CTs_habitat2
}
\description{
Example of second data of coordinates and covariates of each camera trap.
}
\keyword{datasets}
