% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTs_Operativity.R
\docType{data}
\name{CTs_Operativity}
\alias{CTs_Operativity}
\title{Example of data of camera traps (CTs) operativity}
\format{
\code{CTs_Operativity}
\describe{
\item{X}{X-axes coordinates}
\item{Y}{Y-axes coordinates}
\item{Station}{name camera trap}
\item{Fecha_colocacion}{The first day the camera trap was installed}
\item{Fecha_retiro}{The last day that the camera trap was removed}
\item{Problem1_from}{First day that the camera trap stopped working due to some problem}
\item{Problem1_to}{Last day the camera trap stopped working due to some problem}
}
}
\usage{
CTs_Operativity
}
\description{
These data.frame was previously processed with the \code{camtrapR} to obtain a \code{.csv} file.
}
\keyword{datasets}
