% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTs_Spp.R
\docType{data}
\name{CTs_Spp}
\alias{CTs_Spp}
\title{Example of data of mammal species in camera traps.}
\format{
\code{CTs_Spp}
Data frame with 90 rows and 6 variables:
\describe{
\item{Station}{name camera trap}
\item{Species}{Species photographed}
\item{DateTimeOriginal}{Date and time format}
\item{Date}{Date of photo}
\item{Time}{Time of photo}
\item{HierarchicalSubject}{Species and, in some cases, information of category of the animal in the photo as sex, age, identified individual, and other}
}
}
\usage{
CTs_Spp
}
\description{
A data set of mammal species photographed on camera traps in a locality of the Biosphere Reserve in Mexico.
}
\details{
These data.frame was previously processed with the \code{camtrapR} to obtain a \code{.csv} file. Unnecessary columns were deleted.
}
\keyword{datasets}
