% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTs_Spp2.R
\docType{data}
\name{CTs_Spp2}
\alias{CTs_Spp2}
\title{Second example of data of mammal species in camera traps.}
\format{
\code{CTs_Spp2}
Data frame with 32 rows and 7 variables:
\describe{
\item{Camera}{Name camera trap}
\item{Species}{Species photographed}
\item{Events}{Number of independet photos}
\item{Effort}{Number of sampling effort}
\item{Location}{Study area}
\item{Year}{Sampling year}
\item{Season}{Sampling season}
}
}
\usage{
CTs_Spp2
}
\description{
A data set of mammal species photographed on camera traps in a locality of the Biosphere Reserve in Mexico.
}
\keyword{datasets}
