% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Spp.R
\docType{data}
\name{data_Spp}
\alias{data_Spp}
\title{Mammals species in camera traps.}
\format{
\code{data_Spp}
Data frame with 90 rows and 6 variables:
\describe{
\item{Station}{name camera trap}
\item{Species}{Species photographed}
\item{DateTimeOriginal}{Date and time format}
\item{Date}{Date of photo}
\item{Time}{Time of photo}
\item{HierarchicalSubject}{Species and, in some cases, information of category of the animal in the photo as sex, age, identified individual and other}
}
}
\usage{
data_Spp
}
\description{
A data from mammal species photographed on camera traps in a locality of the Biosphere Reserve in Mexico.
}
\details{
These data.frame was previously processed with the \code{camtrapR} to obtained a \code{.csv} file. Unnecesary columns were delete.
}
\keyword{datasets}
