% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFormatDF.R
\name{dataFormatDF}
\alias{dataFormatDF}
\title{Function \code{dataFormatDF}}
\usage{
dataFormatDF(dframe, data_CT, Problem)
}
\arguments{
\item{dframe}{Data of species}

\item{data_CT}{Data of camera traps}

\item{Problem}{Logical. Define possible problem in DataOriginal}
}
\value{
Save a formated data.frame
}
\description{
Function specific to format each data.frame of location, year and season.
}
\examples{
\dontrun{
dataFormatDF(dframe = "data/df/Cb_2012_1.csv", data_CT = CTs, Problem = T)
dataFormatDF(dframe = "data/df/Cb_2012_2.csv", data_CT = CTs, Problem = T)
dataFormatDF(dframe = "data/df/Cb_2013_1.csv", data_CT = CTs, Problem = T)
dataFormatDF(dframe = "data/df/Cb_2013_2.csv", data_CT = CTs, Problem = T)
}

}
\author{
"SMandujanoR"
}
