% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_LYS.R
\name{date_LYS}
\alias{date_LYS}
\title{Format date as "Ymd HM" to "mdY HM"}
\usage{
date_LYS(
  df_Spp,
  Location,
  mdY,
  Jan,
  Feb,
  Mar,
  Apr,
  May,
  Jun,
  Jul,
  Aug,
  Sep,
  Oct,
  Nov,
  Dec,
  Y.init,
  Y.end,
  S.init,
  S.end
)
}
\arguments{
\item{df_Spp}{Data of species}

\item{Location}{Select location}

\item{mdY}{Logical. Define TRUE/FALSE if the DateTimeOriginal column in the data.frame actually is mdY or not}

\item{Jan}{Factor. Define the season (1,2,3...) of this mounth}

\item{Feb}{Factor. Define the season (1,2,3...) of this mounth}

\item{Mar}{Factor. Define the season (1,2,3...) of this mounth}

\item{Apr}{Factor. Define the season (1,2,3...) of this mounth}

\item{May}{Factor. Define the season (1,2,3...) of this mounth}

\item{Jun}{Factor. Define the season (1,2,3...) of this mounth}

\item{Jul}{Factor. Define the season (1,2,3...) of this mounth}

\item{Aug}{Factor. Define the season (1,2,3...) of this mounth}

\item{Sep}{Factor. Define the season (1,2,3...) of this mounth}

\item{Oct}{Factor. Define the season (1,2,3...) of this mounth}

\item{Nov}{Factor. Define the season (1,2,3...) of this mounth}

\item{Dec}{Factor. Define the season (1,2,3...) of this mounth}

\item{Y.init}{First year}

\item{Y.end}{Laste year}

\item{S.init}{First season}

\item{S.end}{Laste season}
}
\value{
This create new columns LYS, the subcarpet df and a new data.frame
}
\description{
Function to 1) format date as "Ymd HM" to "mdY HM", 2) create new columns (Location, Year, Season), and 3) save subsets of data.frames
}
\examples{
\dontrun{
data_Spp <- read.csv("data/CT_Spp.csv", header = T) # Read the original data
data_Spp <- data_Spp[,-1] # If necesesary delete this column
Example 1:
dateSubsets_LYS(df_Spp = data_Spp,
       Location = "CB",
       mdY = T,
       Jan="1",
       Feb="1",
       Mar="1",
       Apr="1",
       May="1",
       Jun="2",
       Jul="2",
       Aug="2",
       Sep="2",
       Oct="2",
       Nov="2",
       Dec="1",
       Y.init = 2012,
       Y.end = 2013,
       S.init = 1,
       S.end = 2)
Example 2: is not necesary modify mdY
dateSubsets_LYS(df_Spp = table2, Location = "CBlanca", mdY = F,
Jan="1", Feb="1", ...
Y.init = 2012, Y.end = 2013,
S.init = 1, S.end = 2)
}

}
\author{
"SMandujanoR and C. García-Vital"
}
