% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eR_facet.R
\name{eR_facet}
\alias{eR_facet}
\title{Function \code{eR_facet}}
\usage{
eR_facet(df, bubSize, fontSize, SaveFolder)
}
\arguments{
\item{df}{Data of species}

\item{bubSize}{Size of the circles}

\item{fontSize}{Size of the fonts}

\item{SaveFolder}{Directory to save plot}
}
\value{
Facet graph eR by species by years and sesons
}
\description{
Function to facet graph eR by species by years and sesons
}
\examples{
\dontrun{
eR_facet(df = datos, bubSize = 7, fontSize = 6, SaveFolder = "Results")
}

}
\author{
"SMandujanoR"
}
