% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eR_glm.R
\name{eR_glm}
\alias{eR_glm}
\title{Function \code{eR_glm}}
\usage{
eR_glm(new.mat, family, SaveFolder)
}
\arguments{
\item{new.mat}{Data of species}

\item{family}{Define "poisson", "quasipoisson"}

\item{SaveFolder}{Directory to save plot}
}
\value{
Table with eR estimation follwing Poisson GLM
}
\description{
Function to estimate eR according with a GLM
}
\examples{
\dontrun{
eR_glm(datos, family = "poisson", SaveFolder = "Results")
}
}
\author{
"SMandujanoR"
}
