% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eR_LYS.R
\name{eR_LYS}
\alias{eR_LYS}
\title{Function \code{eR_LYS}}
\usage{
eR_LYS(df, Factor, Name, bubSize, fontSize, SaveFolder)
}
\arguments{
\item{df}{Data of species}

\item{Factor}{Factor. Select Location, Year or Season}

\item{Name}{Name of Location, Year or Season}

\item{bubSize}{Size of the circles}

\item{fontSize}{Size of the fonts}

\item{SaveFolder}{Directory to save plot}
}
\value{
Graphs of eR for species in different locations and years
}
\description{
Function to graph the eR for species in different locations and years
}
\examples{
\dontrun{
eR_LYS(
  df = datos,
  Factor = datos$Location,
  Name = "Location",
  bubSize = 15,
  fontSize = 12,
  SaveFolder = "Results"
)

eR_LYS(
  df = datos,
  Factor = datos$Year,
  Name = "Year",
  bubSize = 15,
  fontSize = 12,
  SaveFolder = "Results"
)

eR_LYS(
  df = datos,
  Factor = datos$Season,
  Name = "Season",
  bubSize = 15,
  fontSize = 12,
  SaveFolder = "Results"
)
}

}
\author{
"A. Zavaleta & SMandujanoR"
}
