% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eR_matx.R
\name{eR_matx}
\alias{eR_matx}
\title{Create tables of Species-eR, and Species-Events}
\usage{
eR_matx(SaveFolder)
}
\arguments{
\item{SaveFolder}{Directory to save plot}
}
\value{
Tables of Species-eR, and Species-Events
}
\description{
Function to create tables of Species-eR, and Species-Events
}
\examples{
\dontrun{
eR_matx(SaveFolder = "Results")
}
}
\author{
"SMandujanoR"
}
