% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eR_resTab.R
\name{eR_resTab}
\alias{eR_resTab}
\title{Function \code{eR_resTab}}
\usage{
eR_resTab(new.mat, SaveFolder)
}
\arguments{
\item{new.mat}{Data of species}

\item{SaveFolder}{Directory to save plot}
}
\value{
Create Table of results
}
\description{
Function to create a final table
}
\examples{
\dontrun{
eR_resTab(datos, SaveFolder = "Results")
}
}
\author{
"SMandujanoR"
}
