% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eR_test.R
\name{eR_test}
\alias{eR_test}
\title{Statistical test}
\usage{
eR_test(Ymax, SaveFolder)
}
\arguments{
\item{Ymax}{Define maxime value of Y-axis}

\item{SaveFolder}{Directory to save plot}
}
\value{
Anova test and plots of posteriori test
}
\description{
Function to test statistical diferences through parametric ANOVA, and posteriori HSD and Tukey tests
}
\examples{
\dontrun{
eR_test(Ymax = 15, SaveFolder = "Results")
}

}
\author{
"SMandujanoR"
}
