% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map_extract.R
\name{Map_extract}
\alias{Map_extract}
\title{Function \code{Map_extract}}
\usage{
Map_extract(map, CTs, SaveFolder)
}
\arguments{
\item{map}{Shapefile of the study area}

\item{CTs}{Coordinates XY of camera traps}

\item{SaveFolder}{Directory to save plot}
}
\value{
a data.frame with new column named Habt_types
}
\description{
Function to extract information of each camera traps of the raster layer of the study area
}
\examples{
\dontrun{
Map_extract(map = mapProje, CTs = habitat.data, SaveFolder = "Results")
}

}
\author{
"SMandujanoR"
}
