% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCatgRAI.R
\name{MapCatgRAI}
\alias{MapCatgRAI}
\title{Create a map of the RAI selected category}
\usage{
MapCatgRAI(Catg, map, pointSize, colors, SaveFolder)
}
\arguments{
\item{Catg}{Factor. Define the category to map RAI}

\item{map}{shapefile of the study area}

\item{pointSize}{Define the size of the circles}

\item{colors}{Name of your own palete colors}

\item{SaveFolder}{Directory to save results}
}
\value{
Map plot of RAI of selected category as sex, age, other
}
\description{
To create a map of the RAI selected category as sex, age, other
}
\examples{
\dontrun{
MapSexRAI(
  Sex = "Female",
  map = mapProje,
  pointSize = 0.05,
  colors = my_colors,
  SaveFolder = "Cb"
)
}

}
\author{
"SMandujanoR"
}
