% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCTs.R
\name{MapCT}
\alias{MapCT}
\title{Function \code{MapCT}}
\usage{
MapCT(map, CTs, colors, SaveFolder)
}
\arguments{
\item{map}{Shapefile of the study area}

\item{CTs}{Coordinates XY of camera traps}

\item{colors}{Define the colors pallete}

\item{SaveFolder}{Directory to save plot}
}
\value{
plot of camera traps in the study area
}
\description{
Function to map the camera traps in the study area
}
\examples{
\dontrun{
MapCT(map = mapProje, CTs = habitat.data, colors = my_colors, SaveFolder = "Results")
}
}
\author{
"SMandujanoR"
}
