% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapEsriCTs.R
\name{MapEsriCTs}
\alias{MapEsriCTs}
\title{Function \code{MapEsriCTs}}
\usage{
MapEsriCTs(CTs, Proje, reProje)
}
\arguments{
\item{CTs}{Coordinates XY of camera traps}

\item{Proje}{Projection 1 crs}

\item{reProje}{Projection 2 sp::CRS}
}
\value{
map of the camera traps in Esri map projection in the RStudio Viewer
}
\description{
Function to map the camera traps in Esri
}
\examples{
\dontrun{
MapEsriCTs(CTs = habitat.data, Proje = proje2, reProje = "+proj=longlat +datum=WGS84")
}

}
\author{
"G. Andrade-Ponce & SMandujanoR"
}
