% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapEsriRAI.R
\name{MapEsriRAI}
\alias{MapEsriRAI}
\title{RAI Esri map}
\usage{
MapEsriRAI(Location, species, Proje, reProje, pointsize, SaveFolder)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{Proje}{Projection 1 crs}

\item{reProje}{Projection 2 sp::CRS}

\item{pointsize}{Define the size of the circles}

\item{SaveFolder}{Directory to save plot}
}
\value{
Esri map of RAI for selected species
}
\description{
To project RAI of the selected species on Esri.WorldImagery
}
\examples{
\dontrun{
MapEsriRAI(
  Location = "Cb",
  species = "Pec_taj",
  Proje = proje2,
  reProje = "+proj=longlat +datum=WGS84",
  pointsize = 2,
  SaveFolder = "Venado"
)
}

}
\author{
"SMandujanoR"
}
