% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapEventsSpp.R
\name{MapEventsSpp}
\alias{MapEventsSpp}
\title{Plot XY of events}
\usage{
MapEventsSpp(new.mat, especie, parRow, parCol, pointSize, labels, SaveFolder)
}
\arguments{
\item{new.mat}{Data of species}

\item{especie}{Select the species to plot}

\item{parRow}{Number of rows}

\item{parCol}{Number of colummns}

\item{pointSize}{Size of the circles}

\item{labels}{Axes legends}

\item{SaveFolder}{Directory to save plot}
}
\value{
Plot XY of events in camera traps for the selected species
}
\description{
Function to plot XY of events in camera traps for the selected species
}
\examples{
\dontrun{
MapEventsSpp(
  df = datos,
  Species = c("Syl_flo", "Odo_vir", "Can_lat", "Uro_cin"),
  parRow = 2,
  parCol = 2,
  pointSize = 0.5,
  labels = T,
  SaveFolder = "Results"
)
}

}
\author{
"SMandujanoR"
}
