% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapKernelSp.R
\name{MapKernelSp}
\alias{MapKernelSp}
\title{Function \code{MapKernelSp}}
\usage{
MapKernelSp(
  Location,
  species,
  Season,
  map,
  Proje,
  my_colors,
  S.extent,
  UD95,
  UD50,
  pointSize,
  SaveFolder
)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{Season}{Define the season}

\item{map}{shapefile of the study area}

\item{Proje}{Define the projection}

\item{my_colors}{Name of your own palete colors}

\item{S.extent}{Extent}

\item{UD95}{Utilization distribution ("Home range")}

\item{UD50}{Utilization distribution ("Core area")}

\item{pointSize}{Define the size of the circles}

\item{SaveFolder}{Directory to save results}
}
\value{
Estimation, maps and raster of UDs
}
\description{
To create kernel utilization distribution for selected species by season and total (year+season)
}
\examples{
\dontrun{
Season 1:
MapKernelSp(Location =  "Cb",
            species = "Odo_vir",
            Season = 1,
            map = mapProje,
            Proje = proje1,
            my_colors = my_colors,
            S.extent = 1,
            UD95 = 95,
            UD50 = 50,
            pointSize = 0.3,
            SaveFolder = "Venado")

MapKernelSp(Location =  "Cb",
            species = "Odo_vir",
            Season = 2,
            map = mapProje,
            Proje = proje1,
            my_colors = my_colors,
            S.extent = 1,
            UD95 = 95,
            UD50 = 50,
            pointSize = 0.3,
            SaveFolder = "Venado")
}
}
\author{
"SMandujanoR"
}
