% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRAI.R
\name{MapRAI}
\alias{MapRAI}
\title{Function \code{RAIeR}}
\usage{
MapRAI(Location, species, map, pointSize, colors, SaveFolder)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{map}{shapefile of the study area}

\item{pointSize}{Define the size of the circles}

\item{colors}{Name of your own palete colors}

\item{SaveFolder}{Directory to save results}
}
\value{
Map with RAI per camera trap
}
\description{
Function to create a map of the RAI values in each camera trap for the selected species
}
\examples{
\dontrun{
MapRAI(
  Location = "Cb",
  species = "Pec_taj",
  map = mapProje,
  pointSize = 1,
  colors = my_colors,
  SaveFolder = "Pecari"
)
}

}
\author{
"SMandujanoR"
}
