% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAI_facet.R
\name{RAI_facet}
\alias{RAI_facet}
\title{Function \code{RAI_facet}}
\usage{
RAI_facet(Location, species, RAIsize, map, habitats, my_colors, SaveFolder)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{RAIsize}{Define the point size}

\item{map}{shapefile of the study area}

\item{habitats}{Character. Name of habitat o vegetation types level according with the shapefile of the study location}

\item{my_colors}{Name of your own palete colors}

\item{SaveFolder}{Directory to save results}
}
\value{
Facet plots
}
\description{
To comparison RAI by habitat type during each year and season, in a selected location
}
\examples{
\dontrun{
RAI_facet(
  Location = "Cb",
  species = "Pec_taj",
  RAIsize = 40,
  map = mapProje,
  habitats = "tipo_Suelo",
  my_colors = my_colors,
  SaveFolder = "Venado"
)
}

}
\author{
"SMandujanoR"
}
