% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAI_Habitat.R
\name{RAI_Habitat}
\alias{RAI_Habitat}
\title{Function \code{RAI_Habitat}}
\usage{
RAI_Habitat(Location, species, Types, my_colors, SaveFolder)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{Types}{Names. Vegetation or habitat types}

\item{my_colors}{Name of palette colors}

\item{SaveFolder}{Directory to save results}
}
\value{
Linear model and facet plot
}
\description{
To comparison RAI by habitat type during each year and season, in a selected location
}
\examples{
\dontrun{
RAI_Habitat(
  Location = "Cb",
  species = "Odo_vir",
  Types = c(
    "Agriculture", "River", "TDF Neo",
    "TDF Mim", "Saltworks", "Scrub-cras"
  ),
  my_colors = my_colors,
  SaveFolder = "Venado"
)
}

}
\author{
"SMandujanoR"
}
