% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAI_LYS.R
\name{RAI_LYS}
\alias{RAI_LYS}
\title{Graph of RAI per location}
\usage{
RAI_LYS(species, SaveFolder)
}
\arguments{
\item{species}{Select the species}

\item{SaveFolder}{Directory to save results}
}
\value{
Graph RAI by location, year and season
}
\description{
To comparison RAI between locations, years and seasons
}
\examples{
\dontrun{
RAI_LYS(species = "Odo_vir",
     SaveFolder = "Venado")
}

}
\author{
"SMandujanoR"
}
