% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplDesg.R
\name{SamplDesg}
\alias{SamplDesg}
\title{Function \code{SamplDesg}}
\usage{
SamplDesg(map, Proje, n1, n2, CTx, CTy, CTdist, legend, colors, SaveFolder)
}
\arguments{
\item{map}{Shapefile of the study area}

\item{Proje}{Projection 1 crs}

\item{n1}{Number of camera traps in rows}

\item{n2}{Number of camera traps in columns}

\item{CTx}{Initial X-UTM}

\item{CTy}{Initial Y-UTM}

\item{CTdist}{Distance (meters) between cameras}

\item{legend}{Position of legend}

\item{colors}{Define the color palette}

\item{SaveFolder}{Directory to save plot}
}
\value{
Map plot and XY coordinates of each camera trap in the grid
}
\description{
Function to create a camera grid for field sampling desings
}
\examples{
\dontrun{
SamplDesg(
  map = mapProje,
  Proje = proje1,
  n1 = 5,
  n2 = 6,
  CTx = 691000,
  CTy = 2007500,
  CTdist = 1000,
  legend = "bottom",
  colors = my_colors,
  SaveFolder = "Results"
)
}

}
\author{
"SMandujanoR"
}
