% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eR_Gen.R
\name{eR_Gen}
\alias{eR_Gen}
\title{Estimate of encounter rate (eR), naive occupation, and events percent for each species}
\usage{
eR_Gen(new.mat, SaveFolder)
}
\arguments{
\item{new.mat}{Data of species}

\item{SaveFolder}{Directory to save plot}
}
\value{
Table with estimates of encounter rate (eR), naive occupation, and events percent for each species, and plot
}
\description{
Function to estimate of encounter rate (eR), naive occupation, and events percent for each species, and create plots
}
\examples{
\dontrun{
eR_Gen(new.mat = datos, SaveFolder = "Results")
}

}
\author{
"SMandujanoR"
}
