% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapCTs.R
\name{MapCT}
\alias{MapCT}
\title{Map the camera traps}
\usage{
MapCT(map, CTs, mapColor, pointColor, SaveFolder)
}
\arguments{
\item{map}{Shapefile of the study area}

\item{CTs}{Coordinates XY of camera traps}

\item{mapColor}{Select 1 or -1 to define viridis color}

\item{pointColor}{Define color to camera trap point}

\item{SaveFolder}{Directory to save plot}
}
\value{
plot of camera traps in the study area
}
\description{
Function to map the camera traps in the study area
}
\examples{
\dontrun{
MapCT(map = mapProje,
     CTs = habitat.data,
     mapColor = 1,
     pointColor = "black",
     SaveFolder = "Results")
}
}
\author{
"SMandujanoR"
}
