% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapRAI.R
\name{MapRAI}
\alias{MapRAI}
\title{map of the RAI in CTs}
\usage{
MapRAI(Location, species, map, mapColor, pointColor, SaveFolder)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{map}{shapefile of the study area}

\item{mapColor}{Select 1 or -1 to define viridis color}

\item{pointColor}{Define color to camera trap point}

\item{SaveFolder}{Directory to save results}
}
\value{
Map with RAI per camera trap
}
\description{
Function to create a map of the RAI values in each camera trap for the selected species
}
\examples{
\dontrun{
MapRAI(
  Location = "Cb",
  species = "Odo_vir",
  map = mapProje,
  mapColor = 1,
  pointColor = "black",
  SaveFolder = "Venado")
}

}
\author{
"SMandujanoR"
}
