% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAI_facet.R
\name{RAI_facet}
\alias{RAI_facet}
\title{Facet graph for RAI comparison}
\usage{
RAI_facet(Location, species, RAIsize, map, mapColor, pointColor, SaveFolder)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{RAIsize}{Define the point size}

\item{map}{shapefile of the study area}

\item{mapColor}{Select 1 or -1 to define viridis color}

\item{pointColor}{Define color to camera trap point}

\item{SaveFolder}{Directory to save results}
}
\value{
Facet plots
}
\description{
To comparison RAI by habitat type during each year and season, in a selected location
}
\examples{
\dontrun{
RAI_facet(
  Location = "Cb",
  species = "Odo_vir",
  RAIsize = 40,
  map = mapProje,
  mapColor = 1,
  pointColor = "black",
  SaveFolder = "Venado"
)
}

}
\author{
"SMandujanoR"
}
