% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAI_Habitat.R
\name{RAI_Habitat}
\alias{RAI_Habitat}
\title{Graphs of RAI by habitat types}
\usage{
RAI_Habitat(Location, species, fontSize, SaveFolder)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{fontSize}{Size of habitat types names in X-axes}

\item{SaveFolder}{Directory to save results}
}
\value{
Linear model and facet plot
}
\description{
To comparison RAI by habitat types during each year and season, in a selected location
}
\examples{
\dontrun{
RAI_Habitat(
  Location = "Cb",
  species = "Odo_vir",
  fontSize = 8,
  SaveFolder = "Venado")
}

}
\author{
"SMandujanoR"
}
