% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAI_Loc.R
\name{RAI_Loc}
\alias{RAI_Loc}
\title{Graphs of RAI for Year and Season}
\usage{
RAI_Loc(Location, species, Ymax, SaveFolder)
}
\arguments{
\item{Location}{Name location or study area}

\item{species}{Select the species}

\item{Ymax}{Define maxime Y-axes value}

\item{SaveFolder}{Directory to save results}
}
\value{
graphs RAI per Year, Season and Year*Season
}
\description{
To calculate the RAI of the selected species by year, season and season/year, for each studied location
}
\examples{
\dontrun{
RAI_Loc(
  Location = "Cb",
  species = "Odo_vir",
  Ymax = 8,
  SaveFolder = "Venado"
)
}

}
\author{
"SMandujanoR"
}
